+++ processor/src/trace/tests/stack.rs
@@ -2,9 +2,7 @@ use alloc::vec::Vec;
 
 use miden_air::trace::{AUX_TRACE_RAND_ELEMENTS, STACK_AUX_TRACE_OFFSET};
 
-use super::{
-    build_trace_from_ops, rand_array, Felt, FieldElement, Operation, NUM_RAND_ROWS, ONE, ZERO,
-};
+use super::{build_trace_from_ops, rand_array, Felt, FieldElement, Operation, ONE, ZERO};
 use crate::stack::OverflowTableRow;
 
 // CONSTANTS
@@ -92,7 +90,7 @@ fn p1_trace() {
 
     // at this point the table should be empty again, and it should stay empty until the end
     assert_eq!(expected_value, ONE);
-    for i in 13..(p1.len() - NUM_RAND_ROWS) {
+    for i in 13..p1.len() {
         assert_eq!(ONE, p1[i]);
     }
 }