+++ processor/src/trace/utils.rs
@@ -5,7 +5,7 @@ use miden_air::{trace::main_trace::MainTrace, RowIndex};
 #[cfg(test)]
 use vm_core::{utils::ToElements, Operation};
 
-use super::{Felt, FieldElement, NUM_RAND_ROWS};
+use super::{Felt, FieldElement};
 use crate::{chiplets::Chiplets, utils::uninit_vector};
 
 // TRACE FRAGMENT
@@ -124,7 +124,7 @@ impl TraceLenSummary {
 
     /// Returns `trace_len` rounded up to the next power of two.
     pub fn padded_trace_len(&self) -> usize {
-        (self.trace_len() + NUM_RAND_ROWS).next_power_of_two()
+        self.trace_len().next_power_of_two()
     }
 
     /// Returns the percent (0 - 100) of the steps that were added to the trace to pad it to the