+++ verifier/src/lib.rs
@@ -16,7 +16,7 @@ use vm_core::crypto::{
 // EXPORTS
 // ================================================================================================
 pub use vm_core::{chiplets::hasher::Digest, Kernel, ProgramInfo, StackInputs, StackOutputs, Word};
-use winter_verifier::verify as verify_proof;
+use winter_verifier::{crypto::MerkleTree, verify as verify_proof};
 pub use winter_verifier::{AcceptableOptions, VerifierError};
 pub mod math {
     pub use vm_core::{Felt, FieldElement, StarkField};
@@ -69,25 +69,33 @@ pub fn verify(
     match hash_fn {
         HashFunction::Blake3_192 => {
             let opts = AcceptableOptions::OptionSet(vec![ProvingOptions::REGULAR_96_BITS]);
-            verify_proof::<ProcessorAir, Blake3_192, WinterRandomCoin<_>>(proof, pub_inputs, &opts)
+            verify_proof::<ProcessorAir, Blake3_192, WinterRandomCoin<_>, MerkleTree<_>>(
+                proof, pub_inputs, &opts,
+            )
         },
         HashFunction::Blake3_256 => {
             let opts = AcceptableOptions::OptionSet(vec![ProvingOptions::REGULAR_128_BITS]);
-            verify_proof::<ProcessorAir, Blake3_256, WinterRandomCoin<_>>(proof, pub_inputs, &opts)
+            verify_proof::<ProcessorAir, Blake3_256, WinterRandomCoin<_>, MerkleTree<_>>(
+                proof, pub_inputs, &opts,
+            )
         },
         HashFunction::Rpo256 => {
             let opts = AcceptableOptions::OptionSet(vec![
                 ProvingOptions::RECURSIVE_96_BITS,
                 ProvingOptions::RECURSIVE_128_BITS,
             ]);
-            verify_proof::<ProcessorAir, Rpo256, RpoRandomCoin>(proof, pub_inputs, &opts)
+            verify_proof::<ProcessorAir, Rpo256, RpoRandomCoin, MerkleTree<_>>(
+                proof, pub_inputs, &opts,
+            )
         },
         HashFunction::Rpx256 => {
             let opts = AcceptableOptions::OptionSet(vec![
                 ProvingOptions::RECURSIVE_96_BITS,
                 ProvingOptions::RECURSIVE_128_BITS,
             ]);
-            verify_proof::<ProcessorAir, Rpx256, RpxRandomCoin>(proof, pub_inputs, &opts)
+            verify_proof::<ProcessorAir, Rpx256, RpxRandomCoin, MerkleTree<_>>(
+                proof, pub_inputs, &opts,
+            )
         },
     }
     .map_err(VerificationError::VerifierError)?;